+++ client/components/TestQueue/queries.js
@@ -92,6 +92,9 @@ export const TEST_QUEUE_CONFLICTS_PAGE_QUERY = gql`
           directory
         }
       }
+      exactAtVersion {
+        name
+      }
       minimumAtVersion {
         name
       }
@@ -111,6 +114,27 @@ export const TEST_QUEUE_CONFLICTS_PAGE_QUERY = gql`
         ...IssueFieldsAll
       }
       conflicts {
+        source {
+          test {
+            id
+            title
+            rowNumber
+          }
+          scenario {
+            id
+            commands {
+              id
+              text
+              atOperatingMode
+            }
+          }
+          assertion {
+            id
+            text
+            phrase
+            priority
+          }
+        }
         conflictingResults {
           testPlanRun {
             ...TestPlanRunFields
@@ -134,15 +158,18 @@ export const TEST_QUEUE_CONFLICTS_PAGE_QUERY = gql`
           }
           scenarioResult {
             output
+            untestable
             hasNegativeSideEffect
             negativeSideEffects {
+              id
               text
-              details
               impact
+              details
             }
             assertionResults {
               assertion {
                 text
+                priority
               }
               passed
             }